requirements.txt
flask
flask-sqlalchemy
flask-wtf
flask-socketio
python-dotenv
pandas